requirements.txt
django
djangorestframework
django-filter
django-cors-headers
drf-yasg

